+++ src/main/java/com/hbm/capability/HbmLivingProps.java
@@ -5,8 +5,8 @@ import com.hbm.config.RadiationConfig;
 import com.hbm.handler.threading.PacketThreading;
 import com.hbm.lib.ModDamageSource;
 import com.hbm.main.AdvancementManager;
-import com.hbm.packet.toclient.AuxParticlePacketNT;
 import com.hbm.packet.PacketDispatcher;
+import com.hbm.packet.toclient.AuxParticlePacketNT;
 import com.hbm.packet.toclient.PlayerInformPacketLegacy;
 import net.minecraft.block.Block;
 import net.minecraft.entity.EntityLivingBase;
@@ -36,18 +36,18 @@ public class HbmLivingProps {
     }
 
     /// RADIATION ///
-    public static float getRadiation(EntityLivingBase entity) {
+    public static double getRadiation(EntityLivingBase entity) {
         if (!RadiationConfig.enableContamination) return 0;
         return getData(entity).getRads();
     }
 
-    public static void setRadiation(EntityLivingBase entity, float rad) {
+    public static void setRadiation(EntityLivingBase entity, double rad) {
         if (RadiationConfig.enableContamination) getData(entity).setRads(rad);
     }
 
-    public static void incrementRadiation(EntityLivingBase entity, float rad) {
+    public static void incrementRadiation(EntityLivingBase entity, double rad) {
         if (!RadiationConfig.enableContamination) return;
-        float radiation = getRadiation(entity) + rad;
+        double radiation = getRadiation(entity) + rad;
 
         if (radiation > 25000000)
             radiation = 25000000;
@@ -59,43 +59,43 @@ public class HbmLivingProps {
 
     // Neutron Radiation
 
-    public static float getNeutron(EntityLivingBase entity) {
+    public static double getNeutron(EntityLivingBase entity) {
         return getData(entity).getNeutrons();
     }
 
-    public static void setNeutron(EntityLivingBase entity, float rad) {
+    public static void setNeutron(EntityLivingBase entity, double rad) {
         getData(entity).setNeutrons(rad);
     }
 
 
     /// RAD ENV ///
-    public static float getRadEnv(EntityLivingBase entity) {
+    public static double getRadEnv(EntityLivingBase entity) {
         return getData(entity).getRadsEnv();
     }
 
-    public static void setRadEnv(EntityLivingBase entity, float rad) {
+    public static void setRadEnv(EntityLivingBase entity, double rad) {
         getData(entity).setRadsEnv(rad);
     }
 
     /// RAD BUF ///
-    public static float getRadBuf(EntityLivingBase entity) {
+    public static double getRadBuf(EntityLivingBase entity) {
         return getData(entity).getRadBuf();
     }
 
-    public static void setRadBuf(EntityLivingBase entity, float rad) {
+    public static void setRadBuf(EntityLivingBase entity, double rad) {
         getData(entity).setRadBuf(rad);
     }
 
     /// DIGAMA ///
-    public static float getDigamma(EntityLivingBase entity) {
+    public static double getDigamma(EntityLivingBase entity) {
         return getData(entity).getDigamma();
     }
 
-    public static void setDigamma(EntityLivingBase entity, float digamma) {
+    public static void setDigamma(EntityLivingBase entity, double digamma) {
 
         getData(entity).setDigamma(digamma);
 
-        float healthMod = (float) Math.pow(0.5, digamma) - 1F;
+        double healthMod = Math.pow(0.5, digamma) - 1D;
 
         IAttributeInstance attributeinstance = entity.getAttributeMap().getAttributeInstance(SharedMonsterAttributes.MAX_HEALTH);
 
@@ -110,7 +110,7 @@ public class HbmLivingProps {
             entity.setHealth(entity.getMaxHealth());
         }
 
-        if ((entity.getMaxHealth() <= 0 || digamma >= 10.0F) && entity.isEntityAlive()) {
+        if ((entity.getMaxHealth() <= 0 || digamma >= 10.0D) && entity.isEntityAlive()) {
             entity.setAbsorptionAmount(0);
             entity.attackEntityFrom(ModDamageSource.digamma, 5000000F);
             entity.setHealth(0);
@@ -126,19 +126,19 @@ public class HbmLivingProps {
 
         if (entity instanceof EntityPlayer) {
 
-            float di = getData(entity).getDigamma();
+            double di = getData(entity).getDigamma();
 
-            if (di > 0F)
+            if (di > 0D)
                 AdvancementManager.grantAchievement(((EntityPlayer) entity), AdvancementManager.digammaSee);
-            if (di >= 2F)
+            if (di >= 2D)
                 AdvancementManager.grantAchievement(((EntityPlayer) entity), AdvancementManager.digammaFeel);
-            if (di >= 10F)
+            if (di >= 10D)
                 AdvancementManager.grantAchievement(((EntityPlayer) entity), AdvancementManager.digammaKnow);
         }
     }
 
-    public static void incrementDigamma(EntityLivingBase entity, float digamma) {
-        float dRad = getDigamma(entity) + digamma;
+    public static void incrementDigamma(EntityLivingBase entity, double digamma) {
+        double dRad = getDigamma(entity) + digamma;
 
         if (dRad > 10)
             dRad = 10;
@@ -232,12 +232,12 @@ public class HbmLivingProps {
 
     public static class ContaminationEffect {
 
-        public float maxRad;
+        public double maxRad;
         public int maxTime;
         public int time;
         public boolean ignoreArmor;
 
-        public ContaminationEffect(float rad, int time, boolean ignoreArmor) {
+        public ContaminationEffect(double rad, int time, boolean ignoreArmor) {
             this.maxRad = rad;
             this.maxTime = this.time = time;
             this.ignoreArmor = ignoreArmor;
@@ -245,7 +245,7 @@ public class HbmLivingProps {
 
         public static ContaminationEffect load(NBTTagCompound nbt, int index) {
             NBTTagCompound me = nbt.getCompoundTag("cont_" + index);
-            float maxRad = me.getFloat("maxRad");
+            double maxRad = me.getDouble("maxRad");
             int maxTime = me.getInteger("maxTime");
             int time = me.getInteger("time");
             boolean ignoreArmor = me.getBoolean("ignoreArmor");
@@ -255,13 +255,13 @@ public class HbmLivingProps {
             return effect;
         }
 
-        public float getRad() {
-            return maxRad * ((float) time / (float) maxTime);
+        public double getRad() {
+            return maxRad * ((double) time / (double) maxTime);
         }
 
         public void save(NBTTagCompound nbt, int index) {
             NBTTagCompound me = new NBTTagCompound();
-            me.setFloat("maxRad", this.maxRad);
+            me.setDouble("maxRad", this.maxRad);
             me.setInteger("maxTime", this.maxTime);
             me.setInteger("time", this.time);
             me.setBoolean("ignoreArmor", ignoreArmor);
